Deleted the node_modules folder
Deleted this folder to troubleshoot why angular bootstrap was not downloading properly.

Tried running npm install, but that was giving me errors too.  So I ran 'npm install --force', which appeared to have worked.  Then when I tried to install Angular bootstrap, I got the same errors, so I tried to use the --force tag on that, but even this caused errors to occur....